better structure for auth service project
