+++ auth/src/main/java/com/example/auth_api/config/JwtService.java
@@ -24,7 +24,7 @@ import org.springframework.stereotype.Service;
 public class JwtService {
 
   @Value("${JSON_WEB_TOKEN_SECRET}")
-  private String SECRET_KEY;
+  private String secretKey;
 
   public String extractUsername(String token) {
     return extractClaim(token, Claims::getSubject);
@@ -75,7 +75,7 @@ public class JwtService {
    }
   
   private SecretKey getSignInKey() {
-    byte[] keyBytes = Decoders.BASE64.decode(SECRET_KEY);
+    byte[] keyBytes = Decoders.BASE64.decode(secretKey);
     return Keys.hmacShaKeyFor(keyBytes);
   }
 }
