+++ auth/src/main/java/com/example/auth_api/model/AuthResponse.java
@@ -1,4 +1,4 @@
-package com.example.auth_api.auth;
+package com.example.auth_api.model;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -9,9 +9,8 @@ import lombok.NoArgsConstructor;
 @Builder
 @AllArgsConstructor
 @NoArgsConstructor
-public class AuthRequest {
+public class AuthResponse {
 
-  private String email;
-  private String password;
-  
+  private String token;
+
 }
