+++ auth/src/main/java/com/example/auth_api/service/AuthService.java
@@ -1,4 +1,4 @@
-package com.example.auth_api.auth;
+package com.example.auth_api.service;
 
 import org.springframework.stereotype.Service;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
@@ -9,6 +9,9 @@ import com.example.auth_api.model.User;
 import com.example.auth_api.model.Role;
 import com.example.auth_api.repository.UserRepository;
 import com.example.auth_api.config.JwtService;
+import com.example.auth_api.model.AuthRequest;
+import com.example.auth_api.model.AuthResponse;
+import com.example.auth_api.model.RegisterRequest;
 
 import lombok.RequiredArgsConstructor;
